requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dash==3.0.3
dash-bootstrap-components==2.0.2
ecdsa==0.19.1
fastapi==0.115.12
Flask==3.0.3
google-auth==2.39.0
google-auth-oauthlib==1.2.2
greenlet==3.2.1
gspread==6.2.0
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jose==1.0.0
Mako==1.3.10
MarkupSafe==3.0.2
narwhals==1.36.0
nest-asyncio==1.6.0
numpy==2.2.5
oauthlib==3.2.2
packaging==25.0
pandas==2.2.3
passlib==1.7.4
plotly==6.0.1
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==3.0.6
zipp==3.21.0
